// ui/src/common/colors.module.scss.d.ts
interface Colors {
  white: string;
  black: string;
  darkText: string;
  lightText: string;
  lightGrey: string;
  grey: string;
  darkGrey: string;
  red: string;
  green: string;
  yellow: string;
  orange: string;
  blue: string;
  tooltipBg: string;
  shadowDark: string;
  overlayDark: string;
  hoverLight: string;
  borderLight: string;
  borderSubtle: string;
  surfaceLight: string;
  greenTint: string;
  greenBorder: string;
  orangeGlow: string;
  orangeTint: string;
  orangeBadge: string;
  orangeBorder: string;
  orangeTintHover: string;
  orangeBorderHover: string;
  orangeFocus: string;
  redTint: string;
  redBorder: string;
}

const colors: Colors;

export default colors;